requirements.txt
# 完整虚拟环境依赖清单
# 用于完全隔离部署（不依赖系统Python包）
# 适用于: Armbian 32位, Docker, 或其他需要从源码编译的环境

# Web框架
Flask==2.3.3
Flask-CORS==4.0.0

# 图像处理
Pillow==10.0.0

# 数值计算
numpy==1.24.3

# 二维码解码
pyzbar==0.1.9

# Flask依赖
Werkzeug==2.3.7

# 可选: 截图功能（需要X11）
mss==9.0.1

# 可选: 生产环境WSGI服务器
# gunicorn==20.1.0

# 可选: 环境变量管理
# python-dotenv==1.0.0
